move position agent to separate files
