+++ libosmscout-client-qt/include/osmscout/NavigationModule.h
@@ -28,6 +28,7 @@
 #include <osmscout/navigation/Engine.h>
 #include <osmscout/navigation/Agents.h>
 #include <osmscout/navigation/DataAgent.h>
+#include <osmscout/navigation/PositionAgent.h>
 
 #include <osmscout/ClientQtImportExport.h>
