+++ libosmscout/CMakeLists.txt
@@ -64,6 +64,7 @@ set(HEADER_FILES_ROUTING
 set(HEADER_FILES_NAVIGATION
     include/osmscout/navigation/Agents.h
     include/osmscout/navigation/DataAgent.h
+    include/osmscout/navigation/PositionAgent.h
     include/osmscout/navigation/Engine.h
     include/osmscout/navigation/Navigation.h)
 
@@ -165,6 +166,7 @@ set(SOURCE_FILES
     src/osmscout/routing/MultiDBRoutingState.cpp
     src/osmscout/navigation/Agents.cpp
     src/osmscout/navigation/DataAgent.cpp
+    src/osmscout/navigation/PositionAgent.cpp
     src/osmscout/navigation/Engine.cpp
     src/osmscout/Area.cpp
     src/osmscout/AreaDataFile.cpp
