+++ libosmscout/src/meson.build
@@ -47,6 +47,7 @@ osmscoutSrc = [
             'src/osmscout/routing/MultiDBRoutingState.cpp',
             'src/osmscout/navigation/Agents.cpp',
             'src/osmscout/navigation/DataAgent.cpp',
+            'src/osmscout/navigation/PositionAgent.cpp',
             'src/osmscout/navigation/Engine.cpp',
             'src/osmscout/Area.cpp',
             'src/osmscout/AreaDataFile.cpp',
