+++ libosmscout/src/osmscout/navigation/Agents.cpp
@@ -34,120 +34,6 @@ namespace osmscout {
   {
   }
 
-  PositionChangedMessage::PositionChangedMessage(const Timestamp& timestamp,
-                                                 const GeoCoord& currentPosition,
-                                                 double currentSpeed)
-    : NavigationMessage(timestamp),
-      currentPosition(currentPosition),
-      currentSpeed(currentSpeed)
-  {
-  }
-
-  BearingChangedMessage::BearingChangedMessage(const Timestamp& timestamp)
-    : NavigationMessage(timestamp),
-      hasBearing(false),
-      bearing(0.0)
-  {
-  }
-
-  BearingChangedMessage::BearingChangedMessage(const Timestamp& timestamp,
-                                               double bearing)
-    : NavigationMessage(timestamp),
-      hasBearing(true),
-      bearing(bearing)
-  {
-  }
-
-  PositionAgent::PositionAgent()
-  : hasBearing(false),
-    currentState(State::noCoord)
-  {
-  }
-
-  void PositionAgent::UpdateHistory(const GPSUpdateMessage* message)
-  {
-    previousPosition=currentPosition;
-    previousCheckTime=currentCheckTime;
-
-    currentPosition=message->currentPosition;
-    currentCheckTime=message->timestamp;
-    currentSpeed=message->currentSpeed;
-
-    switch (currentState) {
-    case State::noCoord:
-      currentState=State::oneCoord;
-      break;
-    case State::oneCoord:
-      currentState=State::twoCoords;
-      break;
-    case State::twoCoords:
-      break;
-    }
-  }
-
-  std::list<NavigationMessageRef> PositionAgent::UpdateBearing()
-  {
-    std::list<NavigationMessageRef> result;
-
-    if (currentState==State::twoCoords) {
-      hasBearing=true;
-
-      double newBearing=GetSphericalBearingInitial(previousPosition,currentPosition);
-
-      if (currentBearing!=newBearing) {
-        currentBearing=newBearing;
-
-        result.push_back(std::make_shared<BearingChangedMessage>(currentCheckTime,
-                                                                 currentBearing));
-      }
-    }
-
-    return result;
-  }
-
-  std::list<NavigationMessageRef> PositionAgent::Process(const NavigationMessageRef& message)
-  {
-    std::list<NavigationMessageRef> result;
-
-    if (dynamic_cast<GPSUpdateMessage*>(message.get())!=nullptr) {
-      auto gpsUpdateMessage=dynamic_cast<GPSUpdateMessage*>(message.get());
-
-      UpdateHistory(gpsUpdateMessage);
-
-      auto bearingMessages=UpdateBearing();
-
-      result.insert(result.end(),bearingMessages.begin(),bearingMessages.end());
-
-      result.push_back(std::make_shared<PositionChangedMessage>(gpsUpdateMessage->timestamp,
-                                                                gpsUpdateMessage->currentPosition,
-                                                                gpsUpdateMessage->currentSpeed));
-    }
-    else if (dynamic_cast<TimeTickMessage*>(message.get())!=nullptr) {
-      auto now=message->timestamp;
-
-      if (currentState==State::twoCoords &&
-          hasBearing &&
-          now-currentCheckTime>std::chrono::seconds(5)) {
-        previousPosition=currentPosition;
-        previousCheckTime=currentCheckTime;
-
-        auto timeSpanInHours=std::chrono::duration_cast<std::chrono::hours>(now-currentCheckTime).count();
-        auto distance=osmscout::Distance::Of<osmscout::Kilometer>(currentSpeed*timeSpanInHours);
-
-        currentPosition=previousPosition.Add(currentBearing*180/M_PI,
-                                            distance);
-        currentCheckTime=message->timestamp;
-
-
-        result.push_back(std::make_shared<PositionChangedMessage>(message->timestamp,
-                                                                  currentPosition,
-                                                                  currentSpeed));
-      }
-    }
-
-    return result;
-  }
-
   StreetChangedMessage::StreetChangedMessage(const Timestamp& timestamp,
                                              const std::string& name)
   : NavigationMessage(timestamp),
